requirements.txt
Flask
pytesseract 
Pillow
pypdf